+++ scripts/macros/2024/spells/crownOfMadness.js
@@ -1,0 +1,5 @@
+export let crownOfMadness = {
+    name: 'Crown of Madness',
+    version: '1.2.21',
+    rules: 'modern'
+};